requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.124.0
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.14.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-json-logger==4.0.0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.1
uvicorn==0.38.0
watchdog==6.0.0
Werkzeug==3.1.4
